+++ SIL.Media/FFmpegRunner.cs
@@ -12,42 +12,42 @@ using static System.String;
 namespace SIL.Media
 {
 	///<summary>
-	/// FFMpeg is an open source media processing commandline library. Note that there is
-	/// a nuget package called FFMpegCore that wraps the exe to provide this functionality.
+	/// FFmpeg is an open source media processing commandline library. Note that there is
+	/// a nuget package called FFmpegCore that wraps the exe to provide this functionality.
 	/// It has support for asynchronous processing, and with some poking around we could
 	/// maybe figure out how to use that to provide progress reporting, but it's not clear
 	/// that it would be worth it.
 	///</summary>
 	public class FFmpegRunner
 	{
-		private const string kFFMpegExe = "ffmpeg.exe";
+		private const string kFFmpegExe = "ffmpeg.exe";
 		private const string mp3LameCodecArg = "-acodec libmp3lame";
 
 		/// <summary>
-		/// If your app knows where FFMpeg lives, you can tell us before making any calls.
+		/// If your app knows where FFmpeg lives, you can tell us before making any calls.
 		/// </summary>
 		public static string FFmpegLocation;
 
 		/// <summary>
-		/// Find the path to FFMpeg, and remember it (some apps (like SayMore) call FFMpeg a lot)
+		/// Find the path to FFmpeg, and remember it (some apps (like SayMore) call FFmpeg a lot)
 		/// </summary>
 		/// <returns></returns>
-		internal static string LocateAndRememberFFMpeg()
+		internal static string LocateAndRememberFFmpeg()
 		{
 			if (null != FFmpegLocation) //NO! string.empty means we looked and didn't find: string.IsNullOrEmpty(s_ffmpegLocation))
 				return FFmpegLocation;
-			FFmpegLocation = LocateFFMpeg();
+			FFmpegLocation = LocateFFmpeg() ?? Empty;
 			return FFmpegLocation;
 		}
 
 		/// <summary>
-		/// FFMpeg will typically be distributed with SIL software on Windows or automatically
+		/// FFmpeg will typically be distributed with SIL software on Windows or automatically
 		/// installed via package dependencies on other platforms, but if something wants to
 		/// use this library and work with a version of it that the user downloaded or compiled
 		/// locally, this tries to find where they put it.
 		/// </summary>
 		/// <returns>the path, if found, else null</returns>
-		private static string LocateFFMpeg()
+		private static string LocateFFmpeg()
 		{
 			if (Platform.IsLinux)
 			{
@@ -60,14 +60,14 @@ namespace SIL.Media
 				return null;
 			}
 
-			string withApplicationDirectory = GetPathToBundledFFMpeg();
+			string withApplicationDirectory = GetPathToBundledFFmpeg();
 
-			if (!IsNullOrEmpty(withApplicationDirectory) && File.Exists(withApplicationDirectory))
+			if (withApplicationDirectory != null && File.Exists(withApplicationDirectory))
 				return withApplicationDirectory;
 
-			var fromChoco = MediaInfo.GetFFMpegFolderFromChocoInstall(kFFMpegExe);
+			var fromChoco = MediaInfo.GetFFmpegFolderFromChocoInstall(kFFmpegExe);
 			if (fromChoco != null)
-				return Path.Combine(fromChoco, kFFMpegExe);
+				return Path.Combine(fromChoco, kFFmpegExe);
 
 			var progFileDirs = new List<string> {
 				Environment.GetFolderPath(Environment.SpecialFolder.ProgramFilesX86),
@@ -86,46 +86,49 @@ namespace SIL.Media
 			 */
 
 			// REVIEW: I just followed the instructions in the current version of Audacity for
-			// installing FFMpeg for Audacity and the result is a folder that does not contain
+			// installing FFmpeg for Audacity and the result is a folder that does not contain
 			// ffmpeg.exe. This maybe used to work, but I don't think we'll ever find ffmpeg this
 			// way now.
 			//http://manual.audacityteam.org/index.php?title=FAQ:Installation_and_Plug-Ins#installffmpeg
 			foreach (var path in progFileDirs)
 			{
-				var exePath = (Path.Combine(path, "FFmpeg for Audacity", kFFMpegExe));
+				var exePath = (Path.Combine(path, "FFmpeg for Audacity", kFFmpegExe));
 				if (File.Exists(exePath))
 					return exePath;
 			}
-			return Empty;
+			return null;
 		}
 
-		private static string GetPathToBundledFFMpeg()
+		private static string GetPathToBundledFFmpeg()
 		{
 			try
 			{
-				return FileLocationUtilities.GetFileDistributedWithApplication("ffmpeg", kFFMpegExe);
+				return FileLocationUtilities.GetFileDistributedWithApplication("ffmpeg", kFFmpegExe);
 			}
 			catch (Exception)
 			{
-				return Empty;
+				return null;
 			}
 		}
 
 		///<summary>
 		/// Returns false if it can't find ffmpeg
 		///</summary>
-		public static bool HaveNecessaryComponents => !IsNullOrEmpty(LocateFFMpeg());
+		public static bool HaveNecessaryComponents => LocateFFmpeg() != null;
+
+		private static ExecutionResult NoFFmpeg =>
+			new ExecutionResult { StandardError = "Could not locate FFmpeg" };
 
 		///<summary>
 		/// Returns false if it can't find ffmpeg
 		///</summary>
-		private static bool HaveValidFFMpegOnPath
+		private static bool HaveValidFFmpegOnPath
 		{
 			get
 			{
 				if (Platform.IsWindows)
 				{
-					if (!IsNullOrEmpty(LocateFFMpeg()))
+					if (LocateFFmpeg() != null)
 						return true;
 				}
 
@@ -159,13 +162,11 @@ namespace SIL.Media
 		/// <returns>log of the run</returns>
 		public static ExecutionResult ExtractMp3Audio(string inputPath, string outputPath, int channels, IProgress progress)
 		{
-			if (IsNullOrEmpty(LocateFFMpeg()))
-				return new ExecutionResult { StandardError = "Could not locate FFMpeg" };
+			if (LocateFFmpeg() != null)
+				return NoFFmpeg;
 
 			var arguments = $"-i \"{inputPath}\" -vn {mp3LameCodecArg} -ac {channels} \"{outputPath}\"";
-			var result = RunFFMpeg(arguments, progress);
-
-			progress.WriteVerbose(result.StandardOutput);
+			var result = RunFFmpeg(arguments, progress);
 
 			//hide a meaningless error produced by some versions of liblame
 			if (result.StandardError.Contains("lame: output buffer too small")
@@ -196,14 +197,11 @@ namespace SIL.Media
 		/// <returns>log of the run</returns>
 		public static ExecutionResult ExtractOggAudio(string inputPath, string outputPath, int channels, IProgress progress)
 		{
-			if (IsNullOrEmpty(LocateFFMpeg()))
-				return new ExecutionResult { StandardError = "Could not locate FFMpeg" };
+			if (LocateFFmpeg() != null)
+				return NoFFmpeg;
 
 			var arguments = $"-i \"{inputPath}\" -vn -acodec vorbis -ac {channels} \"{outputPath}\"";
-			progress.WriteMessage("ffmpeg " + arguments);
-			var result = RunFFMpeg(arguments, progress);
-
-			progress.WriteVerbose(result.StandardOutput);
+			var result = RunFFmpeg(arguments, progress);
 
 			//hide a meaningless error produced by some versions of liblame
 			if (result.StandardError.Contains("lame: output buffer too small")
@@ -280,8 +278,8 @@ namespace SIL.Media
 		private static ExecutionResult ExtractAudio(string inputPath, string outputPath,
 			string audioCodec, int sampleRate, int channels, IProgress progress)
 		{
-			if (IsNullOrEmpty(LocateFFMpeg()))
-				return new ExecutionResult { StandardError = "Could not locate FFMpeg" };
+			if (LocateFFmpeg() != null)
+				return NoFFmpeg;
 
 			var sampleRateArg = "";
 			if (sampleRate > 0)
@@ -294,11 +292,7 @@ namespace SIL.Media
 
 			var arguments = $"-i \"{inputPath}\" -vn -acodec {audioCodec} {sampleRateArg} {channelsArg} \"{outputPath}\"";
 
-			progress.WriteMessage("ffmpeg " + arguments);
-
-			var result = RunFFMpeg(arguments, progress);
-
-			progress.WriteVerbose(result.StandardOutput);
+			var result = RunFFmpeg(arguments, progress);
 
 			//hide a meaningless error produced by some versions of liblame
 			if (result.StandardError.Contains("lame: output buffer too small")
@@ -327,15 +321,12 @@ namespace SIL.Media
 		public static ExecutionResult ChangeNumberOfAudioChannels(string inputPath,
 			string outputPath, int channels, IProgress progress)
 		{
-			if (IsNullOrEmpty(LocateFFMpeg()))
-				return new ExecutionResult { StandardError = "Could not locate FFMpeg" };
+			if (LocateFFmpeg() != null)
+				return NoFFmpeg;
 
-			var arguments = Format("-i \"{0}\" -vn -ac {1} \"{2}\"",
-				inputPath, channels, outputPath);
+			var arguments = $"-i \"{inputPath}\" -vn -ac {channels} \"{outputPath}\"";
 
-			var result = RunFFMpeg(arguments, progress);
-
-			progress.WriteVerbose(result.StandardOutput);
+			var result = RunFFmpeg(arguments, progress);
 
 			//hide a meaningless error produced by some versions of liblame
 			if (result.StandardError.Contains("lame: output buffer too small") && File.Exists(outputPath))
@@ -363,17 +354,12 @@ namespace SIL.Media
 		/// <returns>log of the run</returns>
 		public static ExecutionResult MakeLowQualityCompressedAudio(string inputPath, string outputPath, IProgress progress)
 		{
-			if (IsNullOrEmpty(LocateAndRememberFFMpeg()))
-				return new ExecutionResult { StandardError = "Could not locate FFMpeg" };
+			if (IsNullOrEmpty(LocateAndRememberFFmpeg()))
+				return NoFFmpeg;
 
 			var arguments = $"-i \"{inputPath}\" {mp3LameCodecArg} -ac 1 -ar 8000 \"{outputPath}\"";
 
-			progress.WriteMessage("ffmpeg " + arguments);
-
-			var result = RunFFMpeg(arguments, progress);
-
-			progress.WriteVerbose(result.StandardOutput);
-
+			var result = RunFFmpeg(arguments, progress);
 
 			//hide a meaningless error produced by some versions of liblame
 			if (result.StandardError.Contains("lame: output buffer too small")
@@ -406,8 +392,8 @@ namespace SIL.Media
 		/// <returns>log of the run</returns>
 		public static ExecutionResult MakeLowQualitySmallVideo(string inputPath, string outputPath, int maxSeconds, IProgress progress)
 		{
-			if (IsNullOrEmpty(LocateAndRememberFFMpeg()))
-				return new ExecutionResult { StandardError = "Could not locate FFMpeg" };
+			if (IsNullOrEmpty(LocateAndRememberFFmpeg()))
+				return NoFFmpeg;
 
 			// isn't working: var arguments = "-i \"" + inputPath + "\" -vcodec mpeg4 -s 160x120 -b 800  -acodec libmp3lame -ar 22050 -ab 32k -ac 1 \"" + outputPath + "\"";
 			var arguments = $"-i \"{inputPath}\" -vcodec mpeg4 -s 160x120 -b 800 " +
@@ -416,11 +402,7 @@ namespace SIL.Media
 				arguments += $" -t {maxSeconds} ";
 			arguments += $"\"{outputPath}\"";
 
-			progress.WriteMessage("ffmpeg " + arguments);
-
-			var result = RunFFMpeg(arguments, progress);
-
-			progress.WriteVerbose(result.StandardOutput);
+			var result = RunFFmpeg(arguments, progress);
 
 			//hide a meaningless error produced by some versions of liblame
 			if (result.StandardError.Contains("lame: output buffer too small")
@@ -449,29 +431,32 @@ namespace SIL.Media
 		[PublicAPI]
 		public static ExecutionResult MakeLowQualitySmallPicture(string inputPath, string outputPath, IProgress progress)
 		{
-			if (IsNullOrEmpty(LocateAndRememberFFMpeg()))
-				return new ExecutionResult { StandardError = "Could not locate FFMpeg" };
+			if (IsNullOrEmpty(LocateAndRememberFFmpeg()))
+				return NoFFmpeg;
 
-			//enhance: how to lower the quality?
+			// ENHANCE: how to lower the quality?
 
 			var arguments = $"-i \"{inputPath}\" -f image2  -s 176x144 \"{outputPath}\"";
 
-			progress.WriteMessage("ffmpeg " + arguments);
+			var result = RunFFmpeg(arguments, progress);
 
-			var result = RunFFMpeg(arguments, progress);
-
-			progress.WriteVerbose(result.StandardOutput);
 			if (result.StandardError.ToLower().Contains("error")) //ffmpeg always outputs config info to standarderror
 				progress.WriteError(result.StandardError);
 
 			return result;
 		}
 
-		private static ExecutionResult RunFFMpeg(string arguments, IProgress progress)
+		private static ExecutionResult RunFFmpeg(string arguments, IProgress progress)
 		{
+			progress.WriteMessage("ffmpeg " + arguments);
+
 			const int timeout = 600; // 60 * 10 = 10 minutes
-			return CommandLineRunner.Run(LocateAndRememberFFMpeg(), arguments,
+			var result = CommandLineRunner.Run(LocateAndRememberFFmpeg(), arguments,
 				Environment.CurrentDirectory, timeout, progress);
+
+			progress.WriteVerbose(result.StandardOutput);
+
+			return result;
 		}
 	}
 }